requirements.txt
# Core dependencies
torch==2.1.0
torchaudio==2.1.0
numpy==1.26.0
librosa==0.10.1
matplotlib==3.8.0
tqdm==4.66.0
jiwer==3.0.1
cmaes==0.11.1

# Whisper from GitHub
git+https://github.com/openai/whisper.git
# Other speech to text models
assemblyai==0.40.2
speechmatics-python==3.0.3
